Individual queries for Item Stats written, 3 more to complete
